requirements.txt
anyio==3.5.0
asgiref==3.4.1
cassandra-driver==3.25.0
click==8.0.3
colorama==0.4.4
fastapi==0.71.0
geomet==0.2.1.post1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
numpy==1.22.0
pydantic==1.9.0
python-dotenv==0.19.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0
